requirements.txt
SpeechRecognition
pyaudio
pyttsx3
python-dotenv
